Replace singleThreadExecutorService with CurrentThreadExecutor.
adal4j's AuthenticationContext requires an Executor service on which to
perform authentication calls.  However, this library makes no use of
that possible concurrency, so there is no reason to create an expensive
new OS level thread for each ADALClientWrapper instance.

It also had been cleaning up the ExecutorService in a `finalize` call,
likely thinking that Java's `finalize` works the way it does in C#.  In
Java, however, `finalize` is less useful (and is deprecated in Java 9),
and as written the code would hold on to the JVM, preventing exit, until
that `finalize` was called (generally two GC sessions later, an
indeterminate amount of time.)  I suspect the code would leak OS threads
as well.